% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_cubist.R
\name{gr_cubist}
\alias{gr_cubist}
\title{Cubist Model}
\usage{
gr_cubist(mf, variable, predvars = NULL)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}
}
\value{
an object of class \code{cubist} with attribute \code{"grundo_model"} \code{"Cubist::cubist"}
}
\description{
Wrapper method for fitting a \code{Cubist::cubist()} model on standard model frame.
}
