% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_gaupro.R
\name{gr_gaupro}
\alias{gr_gaupro}
\title{Gaussian Process Regression (\code{GauPro})}
\usage{
gr_gaupro(mf, variable, predvars = NULL)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}
}
\value{
an object of class \code{GauPro} with attribute \code{"grundo_model"} \code{"gstat::gstat (rkriging)"}
}
\description{
Wrapper method for fitting a \code{GauPro::GauPro()} Gaussian process model
}
