% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_grf.R
\name{gr_grfmrf}
\alias{gr_grfmrf}
\title{Multiple Regression Forest Model (\code{grf})}
\usage{
gr_grfmrf(mf, variable, predvars = NULL, ...)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}

\item{...}{Additional arguments to \code{\link[grf:multi_regression_forest]{grf::multi_regression_forest()}}}
}
\value{
an object of class \code{multi_regression_forest} with attribute \code{"grundo_model"} \code{"grf::multi_regression_forest"}
}
\description{
Wrapper method for fitting a \code{grf::multi_regression_forest()} model on standard model frame.
}
