% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_grf.R
\name{gr_grfqf}
\alias{gr_grfqf}
\title{Quantile Forest Model (\code{grf})}
\usage{
gr_grfqf(mf, variable, predvars = NULL, ...)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}

\item{...}{Additional arguments to \code{\link[grf:quantile_forest]{grf::quantile_forest()}}}
}
\value{
an object of class \code{quantile_forest} with attribute \code{"grundo_model"} \code{"grf::quantile_forest"}
}
\description{
Wrapper method for fitting a \code{grf::quantile_forest()} model on standard model frame.
}
