% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_grf.R
\name{gr_grfrf}
\alias{gr_grfrf}
\title{Regression Forest Model (\code{grf})}
\usage{
gr_grfrf(mf, variable, predvars = NULL, ...)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}

\item{...}{Additional arguments to \code{\link[grf:regression_forest]{grf::regression_forest()}}}
}
\value{
an object of class \code{regression_forest} with attribute \code{"grundo_model"} \code{"grf::regression_forest"}
}
\description{
Wrapper method for fitting a \code{grf::regression_forest()} model on standard model frame.
}
