% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_knnreg.R
\name{gr_knnreg}
\alias{gr_knnreg}
\title{KNN Regression Model (\code{caret})}
\usage{
gr_knnreg(mf, variable, predvars = NULL, k = 5, ...)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}

\item{k}{integer Passed to \code{\link[caret:knnreg]{caret::knnreg()}}}

\item{...}{Additional arguments to \code{\link[caret:knnreg]{caret::knnreg()}}}
}
\value{
an object of class \code{knnreg} with attribute \code{"grundo_model"} \code{"caret::knnreg"}
}
\description{
Wrapper method for fitting a \code{caret::knnreg()} model on standard model frame.
}
