% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_mlreg.R
\name{gr_mlreg}
\alias{gr_mlreg}
\title{Multiple Linear Regression Model (\code{stats})}
\usage{
gr_mlreg(mf, variable, predvars = NULL, step = FALSE, ...)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}

\item{step}{Select the "best" based on AIC step-wise using \code{stats::step()} on standard model frame. Default: \code{FALSE}}

\item{...}{Additional arguments to \code{\link[FNN:knn.reg]{FNN::knn.reg()}}}
}
\value{
an object of class \code{lm} with attribute \code{"grundo_model"} \code{"stats::lm"}
}
\description{
Wrapper method for fitting a \code{stats::lm()} model on standard model frame.
}
\seealso{
\code{\link[=gr_stpreg]{gr_stpreg()}}
}
