% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_ranger.R
\name{gr_ranger}
\alias{gr_ranger}
\title{Random Forest Model (\code{ranger})}
\usage{
gr_ranger(mf, variable, predvars = NULL)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}
}
\value{
an object of class \code{ranger} with attribute \code{"grundo_model"} \code{"ranger::ranger"}
}
\description{
Wrapper method for fitting a \code{ranger::ranger()} model on standard model frame.
}
