% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_rkrigi.R
\name{gr_rkrigi}
\alias{gr_rkrigi}
\title{Regression Kriging (\code{gstat})}
\usage{
gr_rkrigi(
  mf,
  variable,
  predvars = NULL,
  vgm_model = gstat::vgm(),
  coords = c("x", "y")
)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}

\item{vgm_model}{Variogram model selection choices, see \code{\link[gstat:vgm]{gstat::vgm()}} for details}

\item{coords}{Column names in \code{mf} containing spatial coordinates. Default: \code{c("x", "y")}}
}
\value{
an object of class \code{gstat} with attribute \code{"grundo_model"} \code{"gstat::gstat (rkriging)"}
}
\description{
Wrapper method for fitting a \code{gstat::gstat()} regression kriging model
}
