% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_stpreg.R
\name{gr_stpreg}
\alias{gr_stpreg}
\title{Step-wise Regression (\code{stats})}
\usage{
gr_stpreg(mf, variable, predvars = NULL)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}
}
\value{
an object of class \code{lm} with attribute \code{"grundo_model"} \code{"stats::lm"}
}
\description{
Wrapper method for fitting a \code{stats::lm()} model, selecting the "best" based on AIC step-wise using \code{stats::step()} on standard model frame.
}
