% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gr_svm.R
\name{gr_svm}
\alias{gr_svm}
\title{Support Vector Machine/Regression Model (\code{e1071})}
\usage{
gr_svm(mf, variable, predvars = NULL)
}
\arguments{
\item{mf}{\code{data.frame} containing predictors and Observation Column}

\item{variable}{\code{character} Observation Column name}

\item{predvars}{Optional: subset of predictor columns from \code{mf} to use in model fitting}
}
\value{
an object of class \code{"svm.formula"}, \code{"svm"} with attribute \code{"grundo_model"} \code{"e1071::svm"}
}
\description{
Wrapper method for fitting a \code{e1071::svm()} model on standard model frame.
}
