% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grundo-package.R, R/grundo.R
\docType{package}
\name{grundo-package}
\alias{grundo}
\alias{grundo-package}
\title{grundo: Digital Soil Mapping (DSM) From The Ground Up!}
\usage{
grundo(predictors, samples, variable, model, vgm_model = gstat::vgm(), ...)
}
\arguments{
\item{predictors}{\code{SpatRaster} (or coercible to one)}

\item{samples}{\code{SpatVector} (or coercible to one)}

\item{variable}{\code{character}. Observed variable name (in \code{samples})}

\item{model}{Primary model selection choice. Includes: \code{"cubist"} (default), \code{"ranger"} ...}

\item{vgm_model}{Variogram model selection choices for ordinary kriging of primary model residuals, see \code{\link[gstat:vgm]{gstat::vgm()}} for details}

\item{...}{Additional arguments passed to \verb{[gstat::gstat()]}}
}
\value{
an object of class \code{"grundo"}
}
\description{
Simplified utilities for creating gridded maps of soils, with a focus on tools for fitting and evaluating models that predict soil properties and classes.

grundo mundo
}
\keyword{internal}
