% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grundo.R
\name{variogram,SpatVector-method}
\alias{variogram,SpatVector-method}
\title{Variogram Method for SpatVector Objects}
\usage{
\S4method{variogram}{SpatVector}(object, formula, data = as.data.frame(object), ...)
}
\arguments{
\item{object}{\code{SpatVector}}

\item{formula}{\code{formula}, specifying the dependent variable and possible covariates}

\item{data}{Optional: A \code{data.frame} containing variables and regressors}

\item{...}{Additional arguments passed to \code{gstat:::variogram.default()}}
}
\value{
\code{data.frame} result of \code{\link[gstat:variogram]{gstat::variogram()}}
}
\description{
This function uses the \code{gstat::variogram()} generic to define a \verb{variogram(<SpatVector>, <formula>, ...)} method.
}
\details{
This method does not require that the \code{SpatVector} input first be converted to \code{sf} object, preventing dependency on the {sf} namespace being loaded.
}
