% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grundo.R
\name{vgm_models}
\alias{vgm_models}
\title{Variogram Model Types}
\usage{
vgm_models()
}
\value{
character. Vector of variogram model type abbreviations from \code{\link[gstat:vgm]{gstat::vgm()}}
}
\description{
Get choices for variogram model types for use with \code{grundo()} \code{vgm_model} argument.
}
\seealso{
\code{\link[=grundo]{grundo()}}
}
